customer-support-pipeline/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai>=1.30.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
notion-client==2.2.1
requests==2.31.0
APScheduler==3.10.4
gunicorn==21.2.0